CA-425 Avoid race condition, rely on bulkIOs knowledge of active streams.
Checking the stream processors to see if we have one running results in a race
condidtion if a stream is stopped and a new one started directly after. Instead
we can simply ask bulkIO how many streams are active. We can also disable streams
that are added while we are processing.